pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.0.2",
    "langchain-core>=1.0.0",
    "langchain-experimental>=0.0.42",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.1",
    "pre-commit>=4.3.0",
    "python-dotenv>=1.1.1",
]

[tool.ruff]
lint.ignore = ["F401", "E402", "F841"]
